feat: customize currencies select search
